[Geo] Refactor GeoShapeQueryBuilder
Refactors GeoShapeQueryBuilder to derive from a new AbstractShapeQueryBuilder that provides common parsing and build logic for spatial shapes. This will allow development of custom shape queries by extending AbstractShapeQueryBuilder preventing duplication of common shape query logic.
